Bluetooth: controller: Support BT 5.0 feature set bit fields
Added support in the Controller for increased feature set
bit fields in Bluetooth 5.0. Cached feature set in the
connection context is increased to a uint32_t value to
accommodate 17 feature bits.

+++ subsys/bluetooth/controller/ll_sw/ctrl_internal.h
@@ -129,7 +129,7 @@ struct connection {
 		} encryption;
 	} llcp;
 
-	uint8_t llcp_features;
+	uint32_t llcp_features;
 
 	struct {
 		uint8_t tx:1;
